+++ lib/l10n/app_localizations_en.dart
@@ -6618,6 +6618,9 @@ class AppLocalizationsEn extends AppLocalizations {
   String get linkPhraseSupersedesPrimary => 'Supersedes';
 
   @override
+  String get linkTaskButton => 'Link';
+
+  @override
   String get linkedTaskImageBadge => 'From linked task';
 
   @override