+++ lib/l10n/app_localizations_nl.dart
@@ -6671,6 +6671,9 @@ class AppLocalizationsNl extends AppLocalizations {
   String get linkPhraseSupersedesPrimary => 'Supersedes';
 
   @override
+  String get linkTaskButton => 'Link';
+
+  @override
   String get linkedTaskImageBadge => 'Van verbonden taak';
 
   @override