+++ lib/l10n/app_localizations_sv.dart
@@ -6653,6 +6653,9 @@ class AppLocalizationsSv extends AppLocalizations {
   String get linkPhraseSupersedesPrimary => 'Supersedes';
 
   @override
+  String get linkTaskButton => 'Link';
+
+  @override
   String get linkedTaskImageBadge => 'Från länkad uppgift';
 
   @override